[starnix] Watch for exceptions on each thread
Previously, we watched for syscalls on the process object, but that
makes it difficult to create additional threads in the process. Now, we
watch for syscalls on the thread object, which means we can have a
thread in starnix for each thread in the child process once we create
multiple threads in a child process.

Test: No behavior change.
Change-Id: I2c7aefb15639d2d7051d4918aa79ccbafcfa3d5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522201
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Adam Barth <abarth@google.com>