[cm_fidl_analyzer] Implement ComponentInstanceInterface for analyzer
Defines a ComponentModelForAnalyzer which manages
static representations of component instances, each
of which implements the ComponentInstanceInterface
trait.

A ComponentModelForAnalyzer is currently built from
an existing ComponentTree, but may replace
ComponentTree in a follow-up. For now, the two-step
construction will make it easier to transition
Scrutiny's existing route verifier to the new model.

Bug: 61861

Change-Id: I4d55c4dc38d93aacf0a31241c2ad3540eb74d41c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511778
Reviewed-by: Vardhan Mudunuru <vardhan@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>
Commit-Queue: Laura Peskin <pesk@google.com>