[modular] Fix crash when presenting a pending view for a deleted story
When a story is created before sessionmgr chooses a presentation
protocol, i.e. before the session shell's outgoing directory is
served, the story's view is pended, and the view is presented
later when the protocol is chosen.

If the client deleted the story before sessionmgr connected to
the the presentation protocol, sessionmgr would try to read a
nonexistent StoryData and crash.

Fixed: 73335
Change-Id: I746b61413d143eaafa85628d47192f8c0bb8074e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520125
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>