[starnix] Implement getppid, getpgid, and getpgrp
Implementing these functions lead to a refactoring of how tasks are
owned. After this CL, all the task-related data-structures are owned
from the "leaves" of the tree, which are the task objects themselves.

To make the ownership of Task objects clear, Task::new now returns a
TaskOwner object, which has a single owner. When that owner drops the
TaskOwner, the Task object is removed from the ThreadGroup and the
PidTable. When all the Tasks in a ThreadGroup have been removed, the
ThreadGroup kills the underlying Zircon process.

Bug: 74723
Change-Id: I43d80794ec5498e5e8ec9461354dc7ffbb9d6fee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521864
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>