[gidl][llcpp] Special case LLCPP byte vector build benchmarks
The existing benchmarks won't compile quickly for sizes >4k
on arm with clang.

This change adds logic that detects simple repeating patterns
and uses memcpy to reduce the number of operations the
compiler needs to deal with. This was proposed previously,
but I was avoiding doing it because it seemed hacky. Despite
the hackiness, I think this will be useful for benchmarking
iovec encoding.

Bug: 66977
Change-Id: I0dbc76f285e6ebafb0305ffb582b8a6e2a55cd7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520360
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>