[bt][gap] Reduce log severity on errors caused by client disconnect.
A client is allowed to disconnect at anytime. It is not an error to
discover that a client has disconnected when attempting to send it
FIDL messages. Log severity reduced to INFO since it could be useful in
understanding exactly what messages bt-gap believes it sent to a client.

Bug: 75806
Test: Log severity change only.
Change-Id: Ifa3e4791e36a4f0c876de97e440abbe8c3b6f2d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522670
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>