[camera] Have controller allocate client buffers
To solve multiple issues with stream clients connecting in arbitrary
order, have the controller allocate buffers for the clients and hand out
tokens to the already allocated collections.

This required changing the FIDL interface between the camera device and
controller to change the direction of buffer allocation. In addition, the
'SetBufferCollection' method now only uses the passed in token to detect
whether the client is interested in buffers, since the token handed back
to each client via 'WatchBufferCollection' will be from the controller.

This also required changing the internal configuration of the controller
pipelines so that the last node had buffer constraints available, with
sufficient buffers to allow for clients to ask for some 'camping'
buffers as well.

In the process this uncovered several bugs in how controller pipeline
nodes were constructed when streams were opened in a different order.

Also change camera-gym to get rid of the 'only open one stream at a time'
constraint and workaround.

Bug: 42241
Test: controller_unittest, camera_device_unittest

Change-Id: Iaf31f83a028cb5750cf8b7752914175d7f106955
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518926
Commit-Queue: Austin Foxley <afoxley@google.com>
Reviewed-by: Ernest Hua <ernesthua@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>
API-Review: Dale Sather <dalesat@google.com>