[fidlc] Move declaration sorting to post-compilation.
Declaration sorting depends on certain information about the
types being sorted, like the nullability of a type, to determine
the dependencies of a type. In the old syntax we have the
information we need before needing to compile any types,
but this is not the case in the new syntax, so as preparation
for compiling "new syntax" types this CL moves declaration
sorting to be after type compilation.

This also updates the type dependency logic to use Types as
the source of information whenever possible, rather than the
TypeConstructor, which allows us to avoid doing things like
doing a string comparison against "request" to see if a type
refers to a request handle (though currently this hard coding
is OK because fidlc fails to shadow built-in types anyways).

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64-asan/exe.unstripped/fidl-compiler

Change-Id: I6a9761fde7d8dd225e350426de55da38f8411fbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518482
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>