[build][image assembly] use ffx assembly plugin to generate vbmeta image
This adds actions to all builds with vbmeta image generation to
additionally use the new ffx plugin to generate a vbmeta image, and then
an additional action to compare that the two files are the same
(that they have identical contents).

This is a preliminary step to swapping which action's output is used by
the latter steps (while retaining the generation of the file by avbtool)
and then a follow-on change will remove the generate of the file by
avbtool entirely.

Bug: 73187
Test: Lots of CQ dry-runs, with in-build verification step

Change-Id: Ia8f23b582cb1676817c6a51a91fc649db65ccac4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519986
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Shai Barack <shayba@google.com>