[gidl] Add test case for vector of strict enums
This was failing in my in-progress CL and is an important case to cover.
In the C walker, primitive vector elements are no-ops while enums
validate the contents of the element. This means that a vector of enum
covers cases where the value is not a valid member that are not caught
elsewhere.

Change-Id: I812de93a11c6e2ee3bbda6fd72f4003e63fb6197
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523100
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>