[setui] Cancel and wait sensor task
Rather than canceling the sensor task directly, this CL adds a oneshot
to trigger cancelation and then waits for the task to finish completing.
This should help avoid a race where we drop in the middle of waiting
for the sensor to respond, which breaks the connection to the sensor
proxy. Also added an error log in the teardown state in case we discover
that the task is still running on teardown, which indicates other
problems with lifecycle management.

Bug: 74576
Tests: fx test -o setui_service_tests
MULTIPLY: knowledge_card_test: 50
Change-Id: I8adf5796e25c15528b4d8ddea2f02c7f4ab5a400
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522520
Commit-Queue: Paul Faria <paulfaria@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>