[api] Clarifying 'default' value on tables
Using the term 'default' when referring to table fields is ambiguous as
it is unclear who of the sender or receiver is responsible for setting
the default. Instead, clarifying that the receiver is the one doing the
interpretation and considering absent fields as some value.

Bug: 7889
Change-Id: Ia148f04aebc09f33e1bac5f6bd036b0efb91073b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523123
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Pascal-Louis Perez <pascallouis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>