[mdns] fix reader bug - hang on name compression loop
This CL fixes a bug in which DNS message parsing can go into an
infinite loop due to a loop in name compression. That is, if a name
compression reference points to the reference itself, the reader
will spin forever. The root cause was a bad comparison (> rather
than >=) in the name reader.

Bug: 62458

Test: Added unit test that fails before fix
Change-Id: I925fa7d5b2892cb2e5fba1acb1271e849fabb6c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523300
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Reviewed-by: Kostya Kortchinsky <kostyak@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>