[build][image assembly] Use the ffx plugin generated vbmeta image
This switches the vbmeta image used in the update package and flash
files to be the one generated by the ffx plugin, and not the one
generated by avbtool.

However, as a safety check, avbtool still generates an image, which is
compared against the one generated by ffx.  This also requires that
avbtool continues to generate the salt used by both images in their
digests (otherwise they wouldn't directly compare).

The next step will remove the use of avbtool.

Bug: 73190
Test: Lots of CQ
Change-Id: I1266192eeef584a853157207632ed38a94010e5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520064
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>