[blobfs] Fix tests in the new pager
Fixes CompressorTest and HealthCheckService tests using the experimental
Blobfs pager. The new pager is more sensitive to threads and exact setup
and shutdown conditions.

This uses the new test setup class to manage the setup and teardown for
the updated tests. THe health check tests also required that the blobs
be closed to avoid an assertion on teardown about leaking (normally the
VFS handles the open/close, and without this set properly the Blob
doesn't know to unregister itself from the pager watcher which caused
the leak. In the current (original) pager, there is another state where
the Blob can be registered with the pager but not kept alive by it.

The health check tests are the first test using the test setup that
requires blobfs be run on a background thread. A new object is set up
to handle this, and the existing setup functionality is moved to allow
it to be shared between threaded and non-threaded tests.

Change-Id: I4409023c11c4d3bad4839cc9e1cab8d6ac9f6f3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523289
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>