[mtrace][perf_mon][x86] Add basic test of the mtrace syscall
Add a basic test of the mtrace syscall, specifically
MTRACE_PERFMON_GET_PROPERTIES. We expect perf_mon to be available
when the processor reports an architectural PMU with a supported
version.

There is other testing of the entire perf_mon subsystem, covering
the user driver, cpuperf providers, and so on; however there is no
focused testing of the mtrace syscall itself.

This test would have avoided a breakage in perf_mon fixed in
https://fuchsia-review.googlesource.com/c/fuchsia/+/513760 ;
it will also enable testing of the mtrace buffer lifecycle in the
future.

Tested:
$ fx set bringup.x64 --with-base //bundles/bringup:tests
  /boot/test/mtrace

Change-Id: If5991bb45f70d20d5b7884b6e36ff13590337d91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521816
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>