[fxfs] Change the skip_list so that it's drop safe
Previously, we had to rely on commit being called before an iterator is
dropped, but Futures can be dropped at any time so that wasn't safe.
This CL restructures things so that we commit changes when we drop,
which has the nice benefit of no longer requiring the commit method.

Change-Id: I811864be9b20970fc851bf16f14517040488a4d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522905
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>