[virtualization] Update termina_integration_tests.cmx.
The current manifest fails to compile, and is missing two key
capabilities required for the guest instances to start.

Update the cmx file, and add the test package to the appropriate allow
lists.

After this change the termina tests are able to launch the guests, but
the tests still do not successfully start.

Change-Id: Ieb19dbdfb4a7d8dc3066f880125a9ae63de7d69e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522901
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>