[pkdir-integration-tests] land framework
In this CL, we land the testing framework for pkgdir-integration-tests
(e.g. the pkg-harness and the rest of the test topology). In a
subsequent CL, we will create a unique test package for the framework to
use (right now, we're using the system image package).

Bug: 75208
Test: fx test -o pkgdir-integration-tests
Change-Id: I670d26d39ed2ed833a46c6c44c75c7aa9bf7873a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522471
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>