[session][ui-input] Create basic input pipeline component
This creates an input pipeline component that only supports media
buttons. This component is intended to replace input handling in root
presenter. It can be run alongside root presenter, or eventually
scene manager, one of which is required to set up a scene.

Touch support can't currently be added because it requires both the
input pipeline and root presenter to be bound to the scenic session,
which isn't possible. Once the input pipeline library migrates to
using fuchsia.ui.pointerinjector to inject events into scenic
(fxb/75817), this will no longer be an issue.

An integration test for this change can be added following updates
to the existing mediabuttons_listener_test.cc in fxr/520283.

Bug: 74975

Change-Id: I4b2a63864d7457137435e3f991003f497d3e4960
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522671
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Casey Borders <casbor@google.com>
Commit-Queue: Vickie Cheng <vickiecheng@google.com>