[cm][routing] Make check_use routing tests generic
This is the first of several changes making the
component_manager routing tests generic, with the
goal of testing both the CM router and the static
routing analyzer against the same set of test cases.

This CL moves a subset of the CM routing tests into
the routing_test_helpers library. The tests moved in
this CL are those that only need the check_use()
method on a test model.

Bug: 61861
Change-Id: Ic3411f49c127bc9a21f5c5c6df2c9fdf67788158
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523153
Commit-Queue: Laura Peskin <pesk@google.com>
Reviewed-by: Gary Bressler <geb@google.com>