[test-stats][devshell] Pipe timeout parameter to target.
The SSH process that is started along with the fx shell command does not
terminate when fx shell itself does. It seems that it remains connected
until the other side terminates or drops the connection.

In addition to possibly leaving dangling tests, this seems to break
tmux to the point that it is not handling input properly. Changing to
ensure that timeout is piped to the target device at least means that
the other side of the connection will be dropped when there is a timeout
(typically meaning the host side terminal is OK after execution since
the behavior seems to manifest when the other end is stuck on a very
long running test).

Change-Id: Ic5ff0012d40d9e0267fcce453424b8eb7c931ce0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524228
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>