[artifactory] Don't list objs to refresh TTLs for if updated recently.
We have this check in the post-processor that updates the TTLs already,
but because there are a lot of duplicates, it needs to do a lot of
unnecessary Get calls without actually needing to update the CustomTimes
on the objects. Although we'd have to keep track of this in two places,
this should help speed up the post-processor.

Bug: 66462
Change-Id: I4c284b7dc498a498a881b89f97a59d364b6efb49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523992
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>