[fxfs] Fix a SkipListLayer race
When searching for a node, there was a tiny window between the last
comparison and construction of the iterator where a node could have been
inserted which could result in returning a key that's less than the
search key.  The fix is simple: keep hold of the node that is used for
the last comparison and return that to construct the iterator.

Change-Id: I1594b2a2abc3a6e4fb6b5f2c9c112a27bca67de1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523540
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Suter <csuter@google.com>