[audio_core] Move coefficient table creation out of filter.h
This is step 3 of a refactor to support fxrev.dev/471145.
In that CL, we need to reference coefficient table creation functions
from a source file that is not part of the Fuchsia build, meaning the
functions need to be defined in a header file that does not reference
any Fuchsia libraries except the C++ standard libraries. A follow-up
will ensure that mixer/coefficient_table.h does not reference Fuchisa
libraries.

I've also deleted the default ctors, which are unused.

Change-Id: I9181cbbda3651d05e68aa4338b92747079caabae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512155
Commit-Queue: Tom Bergan <tombergan@google.com>
Fuchsia-Auto-Submit: Tom Bergan <tombergan@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>