[ffx] target list names target not found
This also splits the error code functionality from ffx_bail/ffx_error to
ffx_bail_code/ffx_error_code.  Having the error code at the end of the
parameter list causes problems with the macros that end in $($arg:tt) -
so the error code needs to move to the front of the list.  Also, it
would be nice to allow expressions for error codes.  Both of these
changes made the new macros nessecary.

Fixed: 75829

Change-Id: Idd3b3f39ac77fe01da1e757174cc7ab5399f9038
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523985
Commit-Queue: Matthew Boetger <boetger@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>