[build] Use a single toolchain "libprefix" definition.
The toolchain "libprefix" is used to determine the path of the
dynamic linker/loader, when generating Fuchsia executables, as well
as the installation path of shared libraries.

(For more details see the comment added in BUILDCONFIG.gn in the
patch).

This value only depends on the toolchain_variant fields, and was
re-computed in multiple places in our build rules. This CL simply
ensures it is only computed once in //build/config/BUILDCONFIG.gn
as toolchain_variant.libprefix, and converts all uses to refer to
this variable, with one exception (in sysroot_entries.gni below).

This simplifies our build rules a little, and provides a better
way to document this variable/value.

Apart from that, this should not change the build in any way,
and cherry-picking the CL should result in a no-op Ninja build
(after an automatic "gn gen" step though).

+ Remove toolchain.libprefix from
  //build/unification/global_variables.gni by replacing its
  uses with ${toolchain_variant.libprefix} instead.

+ Remove an obsolete comment in zircon_toolchain.gni

Bug: 52720
Bug: 61980
Change-Id: Ib1af76a546289acdbbac8b75684743cc82bf2de7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522300
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>