[lowpan-spinel-driver] Lock network data on NCP before mutating it.
Whenever we mutate properties on the NCP that are a part of the
Thread network data, we must first "lock" this data structure by
setting `SPINEL_PROP_THREAD_ALLOW_LOCAL_NET_DATA_CHANGE` before the
changes and clearing it once the changes are complete. This allows
all of the changes to the network data to be applied at once.

The previous code was not doing this, which caused these methods
to always fail with a `BAD_STATE` error. This change corrects this
by ensuring that the thread network data structure is properly locked
before mutating it.

Note that the solution here isn't perfectly ideal, since there is
a chance that if the returned future is dropped in the middle of
the operation then it won't be given an opportunity to clean up
after itself---specifically, it won't have a chance to disarm the
`SPINEL_PROP_THREAD_ALLOW_LOCAL_NET_DATA_CHANGE` property. Some
thought was given on how to potentially avoid this, but no desirable
solution was immediately obvious. To make sure that such a state
doesn't render the interface hobbled until the next reboot, we will
accept the error `ALREADY` when locking the property. This is
acceptable because the intent of the property was only to have a
cue upon which all of the changes had been made, not mutual
exclusion.

Bug: 75959
Change-Id: I842229e8a915018b9afbaa0f1cbca4551ac0c4eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520503
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>