[lowpan-spinel-driver] Allow commissioning sessions to migrate to mesh
The commissioning session was not fully migrating to the mesh network
once the joining device becomes provisioned. This change makes sure
that the entry is removed from the commissioning session list when the
first packet from the joiner is seen coming over the mesh network,
ensuring that it can continue after.

Bug: 523326
Change-Id: If11fdce7f18ff61cd276d2e67644542ecbedf1cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523326
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>