[fxfs] Fix a journal bug
There was a bug that prevented us replaying starting from the middle of
a block that has a reset checksum.  This CL changes things so that the
reset checksum works slightly differently: we pass a different starting
checksum to the writer rather than indicating reset separately.  This
does mean that detecting a reset now means we have to recompute a
checksum for an entire block, but that won't matter.

Change-Id: I2c02184f86ba500e76288055544320cc20553d27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523780
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>