[starnix] Implement PR_SET_VMA
This CL contains an initial implementation of PR_SET_VMA. Starnix now
keeps track of the memory mappings in the child process. When asked to
set a name for a memory region, Starnix both keeps track of the name
locally and sets the name on the underlying zx::Vmo to integrate with
platform memory tools.

Currently, we do not track mappings created by the ELF loader, which
means some of the PR_SET_VMA operations (e.g., that set the name ".bss")
fail. A future CL will need to track those mappings so we can fork the
child process.

Test: test_prctl_set_vma_anon_name
Fixed: 74726
Change-Id: I0da8c5e18fb894b554ddb5f80efba8ec6ed114da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523402
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>