[pci] Implement PciProtocol in kpci
For the PCI migration there will be a period where drivers may need to
bind as composite drivers or as PciProtocolClients. To enable this,
Kernel PCI needs to support proxy devhosts as well as
PciProtocolClients.  This CL pulls the logic out of Rpc methods and
implements it within Pciprotocol methods, which are then called by the
Rpc methods. Rpc methods handle with any extra management
needed for the wire format for proxy communication.

There is no change to how the devices are added, or how they bind in
this CL.

Test: Manually verified via local builds and CQ testing, no automated
testing exists for legacy kPCI.

Bug: 32978
Change-Id: I94996294e89dcf34af7bb41c8f25eaef02fb77bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517000
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>