[bt][lib] Correct generation bug in DetachableMap
In DetachableMap, you can get a DetachableWeak reference to an item, but
then that item could be replaced later.  If you then detached the original
item, it would remove the new item from the Map, even though you didn't
detach a reference to that item.

Add a bool tracking whether the item is in the parent map, and check it
was true before removing it.

Bug: 75375
Test: added cant_detach_replaced_key
Test: fx test bluetooth-crate-tests
Change-Id: I024bfead7fdfd72273ba034257b420dfe163c3d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520371
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>