[bt][init] Remove IOCapability configuration from bt-init
The Control service now ignores requests to set the IOCapabilities
without a PairingDelegate set. bt-init was using the Control API to set
bt-host's IOCapabilities during stack initialization. There is no way
for a PairingDelegate to be set at this point, as the Control & Access
APIs for setting the PairingDelegate are not yet exposed to the rest of
the system. As such, this attempt to configure does not take effect.
Instead, we implemented non-bondable mode. Now, the system is in non-
bondable mode whenever a pairing delegate is not available.

Bug: b/185940657
Test: `fx test bt-init-unittests` to verify this still builds.
      `fx test bt-profile-integration-tests` passed with logs which
      indicate that bt-init starts up successfully. I've also verified
      that these tests pass even with "stale" default.json config files
      which still have the "io-capability" field (i.e. that this field
      is ignored, not rejected, with the new code).
Change-Id: I9cecf7f9799df168b4720fe75d602eb91a0ef154
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521889
Fuchsia-Auto-Submit: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>