[debugger] Find lines with duplicate names.
In one example Clang is encoding the same file in the line table twice,
one in the "." directory with the full path as the name, and one with
the path as the directory with just the file part as the name.

Previously we were assuming a file only appeared once in the table as an
optimization to avoid unnecessary file name computations. This change
just removes that check and looks up all file names.

Bug: 75991
Change-Id: Icc76ba180a3f6284f9ec3c706892d507f5cd74d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524565
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>