[kernel][tests][timer] Make sure timer objects created are canceled before going out of scope
This avoids a rare race where a timer callback in IRQ mode wakes up a
thread on another cpu or releases the other cpu from a spin loop which
then immediately destroys the Timer object. In this case the Timer
destructor will assert that the timer object appears to be active.

TODO: perhaps a more complete solution involving the destructor either
auto-cancelling or even spin waiting for the timer to complete.

Change-Id: I8b29814e584c8014d22ba4c8f076bd513c11c7a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525360
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>