[fidl][rust] Add result tests to external tests
Results use different code generation logic that is currently not
covered by other tests. This CL gets coverage on changes to this
logic.

This currently triggers an otherwise untested failure in:
Ib33413436eb48eaad97683966f0332894c9f3c1a
This failure is thought to be causing the system to be unable to
boot and this test will help debug it.

Change-Id: I1efbb21d09060ef0ac80bfebc211ad33128898c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525261
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>