[component_manager] Fix work_scheduler_integration_test
The work scheduler test assumed that there would be exactly one
connection established to the Worker protocol. This assumption is
incorrect.

Depending on how tasks get scheduled in component manager, it is
possible for the CapabilityReadyNotifier to attempt to open the Worker
protocol (since it is exposed to framework).

Furthermore, the Notifier uses the DESCRIBE flag to open the service
and waits for an OnOpen event. This is a bug.

Also, ServiceFS silently ignores the DESCRIBE flag when serving
a protocol. Hence the Notifier is waiting for an OnOpen event that
will never arrive. This keeps the channel open while no DoWork
requests are sent on it.

This is why the integration test can sometimes hang.

The proposed fix is three-fold:
1. In the integration test, spawn a task for each client connecting
   to the Worker protocol and use mpsc channels to notify the main
   task when work has been scheduled.
2. CapabilityReadyNotifier should not use the DESCRIBE flag when
   opening exposed protocols.
3. Add a guardrail when ServiceFS attempts to serve protocols
   with the DESCRIBE flag. This should never be supported.

This CL is part 1 of the three-fold fix.

Fixed: 74497
Multiply: work_scheduler_integration_test: 1000
Test: fx test work_scheduler_integration_test
Change-Id: I001305ff2e4d1e193e3b9d2d7edf6175f2071520
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525640
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>