[ffx][iquery] Allow Derivative on host
The iquery code uses the derivative crate to derive equality traits.
Currently iquery only runs on the target and therefore this constraint
was okay. However, I am in the process of generalizing the iquery code
to get some of it to also run on the host. This crate is used in some of
that host code so I want to make it available in that environment.

I couldn't find a compelling reason why it was restricted to fuchsia
only in the first place.

Change-Id: I23aa3be396bf20388cf029bd541ac1bee34c0d21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525105
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>