[volume_image]: Fix pave command line.
Length was supposed to be a default value when not specified,
instead of an overriding value. Renamed variable and switched the
behavior.

Also stat is not a reliable way to obtain number of blocks or size of a
device, so moved to last seekable offset.

Test: storage-volume-image-test
Change-Id: I5cdbe12b7992a27c833fcd8f54533b3e0a0fe4d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525443
Fuchsia-Auto-Submit: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nikhil Raveendranathan <nrav@google.com>