[starnix] Flip FileCommon and FileObject
FileObject is renamed to FileImpl and FileCommon is renamed to
FileObject. FileObject is now the generic bits of an FD, and contains a
Box<&FileImpl> for the filesystem-specific bits.

The main reason I want to do this is to make the API exported by file
objects implemented by generic code rather than filesystem-specific
code. If the outer object is a filesystem-specific trait object, it's
harder to add generic code for a filesystem operation that applies to
every filesystem.

Change-Id: Icba02fde2b6285bd5c4dba54551a661db5493d92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525448
Commit-Queue: Theodore Dubois <tbodt@google.com>
Reviewed-by: Adam Barth <abarth@google.com>