[starnix] Create exception channel before starting process
Previously we started the process before creating the exception channel.
This was causing a race between the first syscall and the exception
channel being installed, which caused tests to flake when the first
syscall won the race.

Verified by introducing a sleep that caused tests to fail consistently
prior to fix and verified that they passed after the fix.

Fixed: 75088, 75975
Change-Id: I4e896b4cd74a1bd376dd34047cdfdd502fb5bb7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526340
Reviewed-by: Theodore Dubois <tbodt@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>