[memory] Generate crash report close to OOM
Write a crash report of type `fuchsia-imminent-oom` when the
Imminent-OOM kernel event is signaled. Unlike crash reports that are
generated at Critical, the Imminent-OOM reports are not rate limited -
OOM events are rarer than Critical crossings, and we would like to
gather information about each of them.

Note that not all Imminent-OOM reports might be followed by an actual
OOM, in case the system is somehow able to recover from memory pressure.
The Imminent-OOM reports are still valuable to have in this case, as
they point to atypical memory usage. If we find that a large number of
reports are being generated, we can consider rate limiting them in the
future.

Test: pressure_notifier tests

Bug: 73872
Change-Id: Iaedf98799015975df9a07eb8d0a7614421724020
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513507
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: David Moore <davemoore@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>