[build] Add action for checking readonly vtables
As an alternative suggestion to
https://fuchsia-review.googlesource.com/c/infra/recipes/+/525703, we can
have a build action run at the end of every build that runs a script for
checking that all vtables are in rodata. This action is not used by default
but can be either run explicitly through ninja or used by adding
`check_vtables_in_rodata=true` to args.gn.

We would only like the script to be run on the clang canary
builders as a last line of defense from potentially breaking upstream
changes that could move vtables out of rodata. A followup infra change
will be added to add `check_vtables_in_rodata=true` to canaries as a gn
argument.

Bug: 45314
Change-Id: I17cc6d628054c1331c2dd5a2103116847311e6d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526220
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>